# docs/swagger.yaml
definitions:
...
  entities.Game:
    properties:
      api_url:
        type: string
      available_rtp:
        items:
          type: integer
        type: array
      available_volatility:
        items:
          type: string
        type: array
      available_wager_sets:
        items:
          $ref: '#/definitions/entities.WagerSet'
        type: array
      available_wager_sets_id:
        items:
          type: string
        type: array
      client_url:
        type: string
      currencies:
        items:
          type: string
        type: array
      gamble_double_up:
        type: integer
      id:
        type: string
      is_demo:
        type: boolean
      is_freespins:
        type: boolean
      is_public:
        type: boolean
      is_statistic_shown:
        type: boolean
      jurisdictions:
        items:
          type: string
        type: array
      languages:
        items:
          type: string
        type: array
      name:
        type: string
      online_volatility:
        type: boolean
      organization:
        $ref: '#/definitions/entities.Organization'
      organization_id:
        type: string
      rtp:
        type: integer
      user_locales:
        items:
          type: string
        type: array
      volatility:
        type: string
      wager_set:
        $ref: '#/definitions/entities.WagerSet'
      wager_set_id:
        type: string
    type: object
...
  entities.IntegratorGameWagerSet:
    properties:
      currency:
        type: string
      game:
        $ref: '#/definitions/entities.Game'
      game_id:
        type: string
      organization:
        $ref: '#/definitions/entities.Organization'
      organization_id:
        type: string
      wager_set:
        $ref: '#/definitions/entities.WagerSet'
      wager_set_id:
        type: string
    type: object
  entities.Organization:
    properties:
      api_key:
        type: string
      id:
        type: string
      name:
        type: string
      status:
        type: integer
      type:
        type: string
    type: object
...
  entities.WagerSet:
    properties:
      created_at:
        type: string
      default_wager:
        type: integer
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      organization_id:
        type: string
      updated_at:
        type: string
      wager_levels:
        items:
          type: integer
        type: array
    type: object
...
  requests.CreateWagerSetRequest:
    properties:
      default_wager:
        type: integer
      name:
        type: string
      wager_levels:
        items:
          type: integer
        type: array
    required:
    - default_wager
    - name
    - wager_levels
    type: object
...
  requests.IntegratorGameWagerSetRequest:
    properties:
      currency:
        type: string
      game_id:
        items:
          type: string
        type: array
      wager_set_id:
        type: string
    required:
    - game_id
    type: object
...
  requests.UpdateIntegratorGameWagerSetRequest:
    properties:
      currency:
        type: string
      game_id:
        type: string
      new_currency:
        type: string
      new_wager_set_id:
        type: string
      wager_set_id:
        type: string
    required:
    - game_id
    type: object
  requests.UpdateWagerSetRequest:
    properties:
      default_wager:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      wager_levels:
        items:
          type: integer
        type: array
    required:
    - default_wager
    - is_active
    - name
    - wager_levels
    type: object
...
  response.Response:
    properties:
      data: {}
      meta: {}
      status:
        type: integer
      success:
        type: boolean
    type: object
...
info:
  contact: {}
paths:
...
  /api/organizations/{id}/wager_set:
    delete:
      consumes:
      - application/json
      description: |-
        The provider_id field is required. If you send only the provider_id,
        all games from the provider will be deleted. To delete selected games, fill out the game_id.
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: organization_id
        in: path
        name: id
        required: true
        type: string
      - description: IntegratorGameWagerSetRequest
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.IntegratorGameWagerSetRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - X-Authenticate: []
      summary: delete integrator game wager set.
      tags:
      - organizations
    get:
      consumes:
      - application/json
      description: get integrator games wager sets.
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: organization_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entities.IntegratorGameWagerSet'
                  type: array
              type: object
      security:
      - X-Authenticate: []
      summary: get integrator games wager sets.
      tags:
      - organizations
    post:
      consumes:
      - application/json
      description: |-
        The provider_id field is required. If you send only the provider_id,
        all games from the provider will be added. To add selected games, fill out the game_id.
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: organization_id
        in: path
        name: id
        required: true
        type: string
      - description: IntegratorGameWagerSetRequest
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.IntegratorGameWagerSetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entities.IntegratorGameWagerSet'
                  type: array
              type: object
      security:
      - X-Authenticate: []
      summary: add integrator game wager sets.
      tags:
      - organizations
    put:
      consumes:
      - application/json
      description: The provider_id field is required. If you send only the provider_id,
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: organization_id
        in: path
        name: id
        required: true
        type: string
      - description: UpdateIntegratorGameWagerSetRequest
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateIntegratorGameWagerSetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/entities.IntegratorGameWagerSet'
              type: object
      security:
      - X-Authenticate: []
      summary: update integrator game wager set.
      tags:
      - organizations
...
  /api/wager_set:
    get:
      consumes:
      - application/json
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: rows limit
        in: query
        name: limit
        required: true
        type: integer
      - description: page
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entities.WagerSet'
                  type: array
              type: object
      security:
      - X-Authenticate: []
      summary: Get wager sets.
      tags:
      - wager_set
    post:
      consumes:
      - application/json
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: requests.CreateWagerSetRequest
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.CreateWagerSetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/entities.WagerSet'
              type: object
      security:
      - X-Authenticate: []
      summary: Create wager set.
      tags:
      - wager_set
  /api/wager_set/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: wager_set_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - X-Authenticate: []
      summary: Delete wager set.
      tags:
      - wager_set
    get:
      consumes:
      - application/json
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: wager_set_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/entities.WagerSet'
              type: object
      security:
      - X-Authenticate: []
      summary: Get wager set.
      tags:
      - wager_set
    put:
      consumes:
      - application/json
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: wager_set_id
        in: path
        name: id
        required: true
        type: string
      - description: requests.UpdateWagerSetRequest
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateWagerSetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/entities.WagerSet'
              type: object
      security:
      - X-Authenticate: []
      summary: Update wager set.
      tags:
      - wager_set
swagger: "2.0"